requirements.txt
psutil
discord.py
async-timeout==3.0.1
attrs==18.2.0
beautifulsoup4==4.6.3
chardet==3.0.4
crontab==0.22.4
idna==2.7
idna-ssl==1.1.0
multidict==4.5.1
PyYAML>=4.2b1
websockets==6.0
yarl==1.2.6
